Refactor Js5 instance names
In the future I will give these more descriptive names.

I'm not yet sure if the client class is the most appropriate location,
but it fits with the current heuristic of placing static members in the
class in which they are initialised if they are initialised in an
instance method.