Add original 549/550 cache
This cache was initially released for build 549 on the 12th June 2009.
The same cache was used at the release of build 550 on the 15th June
2009.

It is available at https://archive.openrs2.org/caches/1.

There are (currently) two colliding groups:

 archive_id | group_id |    crc32    | truncated_version | versions |   containers
------------+----------+-------------+-------------------+----------+-----------------
 5          |     2416 |   565636546 |                 2 | {2}      | {95559,202503}
 9          |      480 | -1980972603 |                 6 | {6}      | {141067,605755}

Archive 5 group 2416 is the NPC spawn file for Taverley. The only
difference is the ID/position of the tool leprachaun. Container 202503
uses 7562, which is a Taverley-specific model. Container 95559 uses
3021, which is a generic model. The generic models were replaced with
location-specific models prior to build 550 - see:

https://runescape.wiki/w/Update:Dungeon_maps_and_spam_filter

Archive 9 group 480 is a texture. It only differs by a single bit, which
doesn't produce a difference in the texture noticeable by the user. I
selected container 141067, as 605755 appears in a much smaller number of
caches and first appeared much later (in 2012).