Use newer 550 cache
The 549/550 cache previously used has some rather critical missing
groups, preventing the client from starting up.

This cache is more complete (only 3 missing groups) and is available at
https://archive.openrs2.org/caches/229.

The only caveat is I don't know if this cache is original or not. The
version numbers of the indexes are between the 549/550 and 551 caches,
which I do know are correct, but that's not definitive proof it's
original.

This commit also updates the set of keys.

The collisions and their resolutions remain the same.